Crate pallet_xcm_bridge_hub_router

Pallet that may be used instead of SovereignPaidRemoteExporter in the XCM router configuration. The main thing that the pallet offers is the dynamic message fee, that is computed based on the bridge queues state. It starts exponentially increasing if the queue between this chain and the sibling/child bridge hub is congested.

All other bridge hub queues offer some backpressure mechanisms. So if at least one of all queues is congested, it will eventually lead to the growth of the queue at this chain.

A note on terminology: when we mention the bridge hub here, we mean the chain that has the messages pallet deployed (pallet-bridge-grandpa, pallet-bridge-messages, pallet-xcm-bridge-hub, …). It may be the system bridge hub parachain or any other chain.

Re-exports

pub use weights::WeightInfo;

pub use pallet::*;

Modules

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

weights

Autogenerated weights for pallet_xcm_bridge_hub_router

Structs

BridgeState

Current status of the bridge.

Constants

HARD_MESSAGE_SIZE_LIMIT

Maximal size of the XCM message that may be sent over bridge.

LOG_TARGET

The target that will be used when publishing logs related to this pallet.

MINIMAL_DELIVERY_FEE_FACTOR

Minimal delivery fee factor.

Traits

XcmChannelStatusProvider

XCM channel status provider that may report whether it is congested or not.