pallet_xcm_bridge_hub_router::pallet

Trait Config

pub trait Config<I: 'static = ()>: Config {
    type RuntimeEvent: From<Event<Self, I>> + IsType<<Self as Config>::RuntimeEvent>;
    type WeightInfo: WeightInfo;
    type UniversalLocation: Get<InteriorLocation>;
    type SiblingBridgeHubLocation: Get<Location>;
    type BridgedNetworkId: Get<Option<NetworkId>>;
    type Bridges: ExporterFor;
    type DestinationVersion: GetVersion;
    type BridgeHubOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type ToBridgeHubSender: SendXcm;
    type LocalXcmChannelManager: XcmChannelStatusProvider;
    type ByteFee: Get<u128>;
    type FeeAsset: Get<AssetId>;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type RuntimeEvent: From<Event<Self, I>> + IsType<<Self as Config>::RuntimeEvent>

The overarching event type.

type WeightInfo: WeightInfo

Benchmarks results from runtime we’re plugged into.

type UniversalLocation: Get<InteriorLocation>

Universal location of this runtime.

type SiblingBridgeHubLocation: Get<Location>

Relative location of the supported sibling bridge hub.

type BridgedNetworkId: Get<Option<NetworkId>>

The bridged network that this config is for if specified. Also used for filtering Bridges by BridgedNetworkId. If not specified, allows all networks pass through.

type Bridges: ExporterFor

Configuration for supported bridged networks/locations with bridge location and possible fee. Allows to externalize better control over allowed bridged networks/locations.

type DestinationVersion: GetVersion

Checks the XCM version for the destination.

type BridgeHubOrigin: EnsureOrigin<Self::RuntimeOrigin>

Origin of the sibling bridge hub that is allowed to report bridge status.

type ToBridgeHubSender: SendXcm

Actual message sender (HRMP or DMP) to the sibling bridge hub location.

type LocalXcmChannelManager: XcmChannelStatusProvider

Local XCM channel manager.

type ByteFee: Get<u128>

Additional fee that is paid for every byte of the outbound message.

type FeeAsset: Get<AssetId>

Asset that is used to paid bridge fee.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.